Представления для списка избранных рецептов пользователя вынесены в отдельный пакет
