+++ recipeblog/urls.py
@@ -5,7 +5,7 @@ urlpatterns = [
     path('', views.index, name='home'),
     path('users/', include('packages.users.urls')),
     path('users/<int:pk>', views.show_user_profile, name='user-detail'),
-    path('usersfavs/<int:pk>/', include('usersfavs.urls')),
+    path('usersfavs/<int:pk>/', include('packages.usersfavs.urls')),
     path('usersrecipes/<int:pk>/', include('usersrecipes.urls')),
     path('recipes/', include('packages.recipes.urls')),
     path('recipes/<int:pk>', views.show_recipe_detail, name='recipe-detail'),
